requirements.txt
chainer==1.17.0
filelock==2.0.6
nose==1.3.7
numpy==1.11.2
opencv-python==3.1.0
Pillow==3.4.2
protobuf==3.1.0.post1
PySide==1.2.2
six==1.10.0
